Make image file preview correct size for its container.
